geo-platform/geosentinel/requirements.txt
requests
python-dotenv
tomtom-api-python
torrequest
